remove main function
