+++ Cargo.toml
@@ -12,6 +12,7 @@ name = "oranda"
 minifier = "0.2.2"
 axohtml = "0.4.1"
 ammonia = "3"
+lazy_static = "1.4.0"
 axum = "0.5.17"
 clap = { version = "4", features = ["derive", "help", "usage", "error-context", "wrap_help"] }
 comrak = "0.14"
