+++ src/main.rs
@@ -4,11 +4,10 @@ mod commands;
 mod config;
 mod errors;
 mod site;
+pub mod tests;
 
 use commands::{Build, Serve};
 use errors::*;
-use tokio::runtime::Runtime;
-
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
 struct Cli {
@@ -21,16 +20,6 @@ enum Command {
     Build(Build),
     Serve(Serve),
 }
-
-pub fn initialize_tokio_runtime() -> Runtime {
-    tokio::runtime::Builder::new_multi_thread()
-        .worker_threads(1)
-        .max_blocking_threads(128)
-        .enable_all()
-        .build()
-        .expect("Initializing tokio runtime failed")
-}
-
 fn main() -> Result<()> {
     let runtime = tokio::runtime::Builder::new_multi_thread()
         .worker_threads(1)
